Move draft button to seperate file
